move error handling middleware to its separate file
